projet-ia/requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
PyOpenGL==3.1.7
numpy==1.26.3 